Better separation of logic. P4_ElasticSearch class now extends P4_Class which is now an abstract class. Separate logic for setting WP_Query args for selected filters. Also, separated logic for settings args based on whether they are related to application or search engine.
